+++ bukkit-loader/src/main/resources/resources/internal/resourcepack/assets/minecraft/textures/item/custom/gui/cooking_info.png.mcmeta
@@ -1,0 +1,6 @@
+{
+  "animation": {
+    "interpolate": false,
+    "frametime": 5
+  }
+}
