+++ core/src/main/java/net/momirealms/craftengine/core/plugin/gui/GuiParameters.java
@@ -5,4 +5,6 @@ import net.momirealms.craftengine.core.util.context.ContextKey;
 public class GuiParameters {
     public static final ContextKey<String> MAX_PAGE = ContextKey.of("max_page");
     public static final ContextKey<String> CURRENT_PAGE = ContextKey.of("current_page");
+    public static final ContextKey<String> COOKING_TIME = ContextKey.of("cooking_time");
+    public static final ContextKey<String> COOKING_EXPERIENCE = ContextKey.of("cooking_experience");
 }
