+++ core/src/main/java/net/momirealms/craftengine/core/plugin/gui/category/ItemBrowserManagerImpl.java
@@ -1,6 +1,5 @@
 package net.momirealms.craftengine.core.plugin.gui.category;
 
-import dev.dejvokep.boostedyaml.block.implementation.Section;
 import net.momirealms.craftengine.core.entity.player.Player;
 import net.momirealms.craftengine.core.item.Item;
 import net.momirealms.craftengine.core.item.ItemBuildContext;
@@ -8,7 +7,6 @@ import net.momirealms.craftengine.core.item.ItemKeys;
 import net.momirealms.craftengine.core.item.recipe.*;
 import net.momirealms.craftengine.core.pack.Pack;
 import net.momirealms.craftengine.core.plugin.CraftEngine;
-import net.momirealms.craftengine.core.plugin.config.ConfigManager;
 import net.momirealms.craftengine.core.plugin.gui.Ingredient;
 import net.momirealms.craftengine.core.plugin.gui.*;
 import net.momirealms.craftengine.core.registry.Holder;
@@ -19,8 +17,6 @@ import net.momirealms.craftengine.core.util.context.ContextHolder;
 
 import java.nio.file.Path;
 import java.util.*;
-
-import static java.util.Objects.requireNonNull;
 
 @SuppressWarnings("DuplicatedCode")
 public class ItemBrowserManagerImpl implements ItemBrowserManager {
@@ -124,7 +120,7 @@ public class ItemBrowserManagerImpl implements ItemBrowserManager {
             item.load();
             return new ItemWithAction(item, (element, click) -> {
                 click.cancel();
-                player.playSound(BUTTON_SOUND);
+                player.playSound(Constants.SOUND_CLICK_BUTTON);
                 openCategoryPage(click.clicker(), it.id(), element.gui());
             });
         }).filter(Objects::nonNull).toList();
@@ -158,7 +154,7 @@ public class ItemBrowserManagerImpl implements ItemBrowserManager {
                 .orElseThrow(() -> new GuiElementMissingException("Can't find gui element " + Constants.CATEGORY_BACK)),
                 ((element, click) -> {
                     click.cancel();
-                    player.playSound(RETURN_SOUND, 0.25f, 1);
+                    player.playSound(Constants.SOUND_RETURN_PAGE, 0.25f, 1);
                     parentGui.open(player);
                 }))
         )
@@ -203,7 +199,7 @@ public class ItemBrowserManagerImpl implements ItemBrowserManager {
                     return;
                 }
                 List<Recipe<Object>> inRecipes = this.plugin.recipeManager().getRecipeByResult(it);
-                player.playSound(BUTTON_SOUND);
+                player.playSound(Constants.SOUND_CLICK_BUTTON);
                 if (!inRecipes.isEmpty()) {
                     openRecipePage(c.clicker(), it, e.gui(), inRecipes, 0, 0);
                 } else {
@@ -245,7 +241,7 @@ public class ItemBrowserManagerImpl implements ItemBrowserManager {
         }))
         .addIngredient('^', player.hasPermission(GET_ITEM_PERMISSION) ? GuiElement.constant(this.plugin.itemManager().createWrappedItem(Constants.RECIPE_GET_ITEM, player), (e, c) -> {
             c.cancel();
-            player.playSound(PICKUP_SOUND);
+            player.playSound(Constants.SOUND_PICK_ITEM);
             if (LEFT_CLICK.contains(c.type())) {
                 player.giveItem(this.plugin.itemManager().createWrappedItem(result, player));
             } else if (RIGHT_CLICK.contains(c.type())) {
@@ -258,7 +254,7 @@ public class ItemBrowserManagerImpl implements ItemBrowserManager {
                         .orElseThrow(() -> new GuiElementMissingException("Can't find gui element " + Constants.RECIPE_BACK)),
                 ((element, click) -> {
                     click.cancel();
-                    player.playSound(RETURN_SOUND, 0.25f, 1);
+                    player.playSound(Constants.SOUND_RETURN_PAGE, 0.25f, 1);
                     parentGui.open(player);
                 }))
         );
@@ -322,7 +318,7 @@ public class ItemBrowserManagerImpl implements ItemBrowserManager {
         }))
         .addIngredient('^', player.hasPermission(GET_ITEM_PERMISSION) ? GuiElement.constant(this.plugin.itemManager().createWrappedItem(Constants.RECIPE_GET_ITEM, player), (e, c) -> {
             c.cancel();
-            player.playSound(PICKUP_SOUND);
+            player.playSound(Constants.SOUND_PICK_ITEM);
             if (LEFT_CLICK.contains(c.type())) {
                 player.giveItem(this.plugin.itemManager().createWrappedItem(result, player));
             } else if (RIGHT_CLICK.contains(c.type())) {
@@ -340,7 +336,7 @@ public class ItemBrowserManagerImpl implements ItemBrowserManager {
             }
             List<Recipe<Object>> inRecipes = plugin.recipeManager().getRecipeByResult(e.item().id());
             if (!inRecipes.isEmpty()) {
-                player.playSound(BUTTON_SOUND);
+                player.playSound(Constants.SOUND_CLICK_BUTTON);
                 openRecipePage(player, e.item().id(), e.gui(), inRecipes, 0, depth + 1);
             }
         }))
@@ -349,7 +345,7 @@ public class ItemBrowserManagerImpl implements ItemBrowserManager {
                         .orElseThrow(() -> new GuiElementMissingException("Can't find gui element " + Constants.RECIPE_BACK)),
                 ((element, click) -> {
                     click.cancel();
-                    player.playSound(RETURN_SOUND, 0.25f, 1);
+                    player.playSound(Constants.SOUND_RETURN_PAGE, 0.25f, 1);
                     parentGui.open(player);
                 }))
         )
@@ -362,7 +358,7 @@ public class ItemBrowserManagerImpl implements ItemBrowserManager {
                 .orElseThrow(() -> new GuiElementMissingException("Can't find gui element " + next)), (e, c) -> {
             c.cancel();
             if (index + 1 < recipes.size()) {
-                player.playSound(PAGE_SOUND, 0.25f, 1);
+                player.playSound(Constants.SOUND_CHANGE_PAGE, 0.25f, 1);
                 openRecipePage(player, result, parentGui, recipes, index + 1, depth);
             }
         }))
@@ -375,7 +371,7 @@ public class ItemBrowserManagerImpl implements ItemBrowserManager {
                 .orElseThrow(() -> new GuiElementMissingException("Can't find gui element " + previous)), (e, c) -> {
             c.cancel();
             if (index > 0) {
-                player.playSound(PAGE_SOUND, 0.25f, 1);
+                player.playSound(Constants.SOUND_CHANGE_PAGE, 0.25f, 1);
                 openRecipePage(player, result, parentGui, recipes, index - 1, depth);
             }
         }));
@@ -406,7 +402,7 @@ public class ItemBrowserManagerImpl implements ItemBrowserManager {
                 "         ",
                 "         ",
                 "  A   X  ",
-                "      ^  ",
+                "  ?   ^  ",
                 "         ",
                 " <  =  > "
         )
@@ -418,9 +414,15 @@ public class ItemBrowserManagerImpl implements ItemBrowserManager {
                 c.setItemOnCursor(item);
             }
         }))
+        .addIngredient('?', GuiElement.constant(this.plugin.itemManager().getCustomItem(Constants.RECIPE_COOKING_INFO)
+                .map(it -> it.buildItem(ItemBuildContext.of(player, ContextHolder.builder()
+                        .withParameter(GuiParameters.COOKING_TIME, String.valueOf(recipe.cookingTime()))
+                        .withParameter(GuiParameters.COOKING_EXPERIENCE, String.valueOf(recipe.experience()))
+                        .build())))
+                .orElseThrow(() -> new GuiElementMissingException("Can't find gui element " + Constants.RECIPE_COOKING_INFO)), (e, c) -> c.cancel()))
         .addIngredient('^', player.hasPermission(GET_ITEM_PERMISSION) ? GuiElement.constant(this.plugin.itemManager().createWrappedItem(Constants.RECIPE_GET_ITEM, player), (e, c) -> {
             c.cancel();
-            player.playSound(PICKUP_SOUND);
+            player.playSound(Constants.SOUND_PICK_ITEM);
             if (LEFT_CLICK.contains(c.type())) {
                 player.giveItem(this.plugin.itemManager().createWrappedItem(result, player));
             } else if (RIGHT_CLICK.contains(c.type())) {
@@ -438,7 +440,7 @@ public class ItemBrowserManagerImpl implements ItemBrowserManager {
             }
             List<Recipe<Object>> inRecipes = plugin.recipeManager().getRecipeByResult(e.item().id());
             if (!inRecipes.isEmpty()) {
-                player.playSound(BUTTON_SOUND);
+                player.playSound(Constants.SOUND_CLICK_BUTTON);
                 openRecipePage(player, e.item().id(), e.gui(), inRecipes, 0, depth + 1);
             }
         }))
@@ -447,7 +449,7 @@ public class ItemBrowserManagerImpl implements ItemBrowserManager {
                         .orElseThrow(() -> new GuiElementMissingException("Can't find gui element " + Constants.RECIPE_BACK)),
                 ((element, click) -> {
                     click.cancel();
-                    player.playSound(RETURN_SOUND, 0.25f, 1);
+                    player.playSound(Constants.SOUND_RETURN_PAGE, 0.25f, 1);
                     parentGui.open(player);
                 }))
         )
@@ -460,7 +462,7 @@ public class ItemBrowserManagerImpl implements ItemBrowserManager {
                 .orElseThrow(() -> new GuiElementMissingException("Can't find gui element " + next)), (e, c) -> {
             c.cancel();
             if (index + 1 < recipes.size()) {
-                player.playSound(PAGE_SOUND, 0.25f, 1);
+                player.playSound(Constants.SOUND_CHANGE_PAGE, 0.25f, 1);
                 openRecipePage(player, result, parentGui, recipes, index + 1, depth);
             }
         }))
@@ -473,7 +475,7 @@ public class ItemBrowserManagerImpl implements ItemBrowserManager {
                 .orElseThrow(() -> new GuiElementMissingException("Can't find gui element " + previous)), (e, c) -> {
             c.cancel();
             if (index > 0) {
-                player.playSound(PAGE_SOUND, 0.25f, 1);
+                player.playSound(Constants.SOUND_CHANGE_PAGE, 0.25f, 1);
                 openRecipePage(player, result, parentGui, recipes, index - 1, depth);
             }
         }));
@@ -524,7 +526,7 @@ public class ItemBrowserManagerImpl implements ItemBrowserManager {
         }))
         .addIngredient('^', player.hasPermission(GET_ITEM_PERMISSION) ? GuiElement.constant(this.plugin.itemManager().createWrappedItem(Constants.RECIPE_GET_ITEM, player), (e, c) -> {
             c.cancel();
-            player.playSound(PICKUP_SOUND);
+            player.playSound(Constants.SOUND_PICK_ITEM);
             if (LEFT_CLICK.contains(c.type())) {
                 player.giveItem(this.plugin.itemManager().createWrappedItem(result, player));
             } else if (RIGHT_CLICK.contains(c.type())) {
@@ -537,7 +539,7 @@ public class ItemBrowserManagerImpl implements ItemBrowserManager {
                         .orElseThrow(() -> new GuiElementMissingException("Can't find gui element " + Constants.RECIPE_BACK)),
                 ((element, click) -> {
                     click.cancel();
-                    player.playSound(RETURN_SOUND, 0.25f, 1);
+                    player.playSound(Constants.SOUND_RETURN_PAGE, 0.25f, 1);
                     parentGui.open(player);
                 }))
         )
@@ -550,7 +552,7 @@ public class ItemBrowserManagerImpl implements ItemBrowserManager {
                 .orElseThrow(() -> new GuiElementMissingException("Can't find gui element " + next)), (e, c) -> {
             c.cancel();
             if (index + 1 < recipes.size()) {
-                player.playSound(PAGE_SOUND, 0.25f, 1);
+                player.playSound(Constants.SOUND_CHANGE_PAGE, 0.25f, 1);
                 openRecipePage(player, result, parentGui, recipes, index + 1, depth);
             }
         }))
@@ -563,7 +565,7 @@ public class ItemBrowserManagerImpl implements ItemBrowserManager {
                 .orElseThrow(() -> new GuiElementMissingException("Can't find gui element " + previous)), (e, c) -> {
             c.cancel();
             if (index > 0) {
-                player.playSound(PAGE_SOUND, 0.25f, 1);
+                player.playSound(Constants.SOUND_CHANGE_PAGE, 0.25f, 1);
                 openRecipePage(player, result, parentGui, recipes, index - 1, depth);
             }
         }));
@@ -594,7 +596,7 @@ public class ItemBrowserManagerImpl implements ItemBrowserManager {
                                 }
                                 List<Recipe<Object>> inRecipes = this.plugin.recipeManager().getRecipeByResult(e.item().id());
                                 if (!inRecipes.isEmpty()) {
-                                    player.playSound(BUTTON_SOUND);
+                                    player.playSound(Constants.SOUND_CLICK_BUTTON);
                                     openRecipePage(player, e.item().id(), e.gui(), inRecipes, 0, depth + 1);
                                 }
                             }));
@@ -625,7 +627,7 @@ public class ItemBrowserManagerImpl implements ItemBrowserManager {
                             }
                             List<Recipe<Object>> inRecipes = this.plugin.recipeManager().getRecipeByResult(e.item().id());
                             if (!inRecipes.isEmpty()) {
-                                player.playSound(BUTTON_SOUND);
+                                player.playSound(Constants.SOUND_CLICK_BUTTON);
                                 openRecipePage(player, e.item().id(), e.gui(), inRecipes, 0, depth + 1);
                             }
                         }));
@@ -648,69 +650,5 @@ public class ItemBrowserManagerImpl implements ItemBrowserManager {
                 .title(AdventureHelper.miniMessage().deserialize(Constants.RECIPE_CRAFTING_TITLE, ItemBuildContext.of(player, ContextHolder.EMPTY).tagResolvers()))
                 .refresh()
                 .open(player);
-    }
-
-    static class Constants {
-        public static String BROWSER_TITLE;
-        public static Key BROWSER_NEXT_PAGE_AVAILABLE;
-        public static Key BROWSER_NEXT_PAGE_BLOCK;
-        public static Key BROWSER_PREVIOUS_PAGE_AVAILABLE;
-        public static Key BROWSER_PREVIOUS_PAGE_BLOCK;
-
-        public static String CATEGORY_TITLE;
-        public static Key CATEGORY_BACK;
-        public static Key CATEGORY_NEXT_PAGE_AVAILABLE;
-        public static Key CATEGORY_NEXT_PAGE_BLOCK;
-        public static Key CATEGORY_PREVIOUS_PAGE_AVAILABLE;
-        public static Key CATEGORY_PREVIOUS_PAGE_BLOCK;
-
-        public static String RECIPE_NONE_TITLE;
-        public static String RECIPE_BLASTING_TITLE;
-        public static String RECIPE_SMELTING_TITLE;
-        public static String RECIPE_SMOKING_TITLE;
-        public static String RECIPE_CAMPFIRE_TITLE;
-        public static String RECIPE_CRAFTING_TITLE;
-        public static String RECIPE_STONECUTTING_TITLE;
-        public static Key RECIPE_BACK;
-        public static Key RECIPE_NEXT_PAGE_AVAILABLE;
-        public static Key RECIPE_NEXT_PAGE_BLOCK;
-        public static Key RECIPE_PREVIOUS_PAGE_AVAILABLE;
-        public static Key RECIPE_PREVIOUS_PAGE_BLOCK;
-        public static Key RECIPE_GET_ITEM;
-
-        public static void load() {
-            Section section = ConfigManager.instance().settings().getSection("gui.browser");
-            if (section == null) return;
-            BROWSER_TITLE = getOrThrow(section, "main.title");
-            BROWSER_NEXT_PAGE_AVAILABLE = Key.of(getOrThrow(section, "main.page-navigation.next.available"));
-            BROWSER_NEXT_PAGE_BLOCK = Key.of(getOrThrow(section, "main.page-navigation.next.not-available"));
-            BROWSER_PREVIOUS_PAGE_AVAILABLE = Key.of(getOrThrow(section, "main.page-navigation.previous.available"));
-            BROWSER_PREVIOUS_PAGE_BLOCK = Key.of(getOrThrow(section, "main.page-navigation.previous.not-available"));
-
-            CATEGORY_TITLE = getOrThrow(section, "category.title");
-            CATEGORY_BACK = Key.of(getOrThrow(section, "category.page-navigation.return"));
-            CATEGORY_NEXT_PAGE_AVAILABLE = Key.of(getOrThrow(section, "category.page-navigation.next.available"));
-            CATEGORY_NEXT_PAGE_BLOCK = Key.of(getOrThrow(section, "category.page-navigation.next.not-available"));
-            CATEGORY_PREVIOUS_PAGE_AVAILABLE = Key.of(getOrThrow(section, "category.page-navigation.previous.available"));
-            CATEGORY_PREVIOUS_PAGE_BLOCK = Key.of(getOrThrow(section, "category.page-navigation.previous.not-available"));
-
-            RECIPE_NONE_TITLE = getOrThrow(section, "recipe.none.title");
-            RECIPE_BLASTING_TITLE = getOrThrow(section, "recipe.blasting.title");
-            RECIPE_SMELTING_TITLE = getOrThrow(section, "recipe.smelting.title");
-            RECIPE_SMOKING_TITLE = getOrThrow(section, "recipe.smoking.title");
-            RECIPE_CAMPFIRE_TITLE = getOrThrow(section, "recipe.campfire.title");
-            RECIPE_CRAFTING_TITLE = getOrThrow(section, "recipe.crafting.title");
-            RECIPE_STONECUTTING_TITLE = getOrThrow(section, "recipe.stonecutting.title");
-            RECIPE_BACK = Key.of(getOrThrow(section, "recipe.page-navigation.return"));
-            RECIPE_NEXT_PAGE_AVAILABLE = Key.of(getOrThrow(section, "recipe.page-navigation.next.available"));
-            RECIPE_NEXT_PAGE_BLOCK = Key.of(getOrThrow(section, "recipe.page-navigation.next.not-available"));
-            RECIPE_PREVIOUS_PAGE_AVAILABLE = Key.of(getOrThrow(section, "recipe.page-navigation.previous.available"));
-            RECIPE_PREVIOUS_PAGE_BLOCK = Key.of(getOrThrow(section, "recipe.page-navigation.previous.not-available"));
-            RECIPE_GET_ITEM = Key.of(getOrThrow(section, "recipe.get-item-icon"));
-        }
-
-        private static String getOrThrow(Section section, String route) {
-            return requireNonNull(section.getString(route), "gui.browser." + route);
-        }
     }
 }
