+++ gradle.properties
@@ -1,7 +1,7 @@
 # Project settings
 # Rule: [major update].[feature update].[bug fix]
-project_version=0.0.24
-config_version=5
+project_version=0.0.25
+config_version=6
 project_group=net.momirealms
 legacy_model_templates_version=1.21.3
 latest_model_templates_version=1.21.4
